+++ server/utils/__tests__/jsonApi.spec.js
@@ -4,6 +4,7 @@ const {
   getSmallTile,
   getCategoryIds,
   buildSecondaryTags,
+  typeFrom,
 } = require('../jsonApi');
 
 const LARGE_TILE = 'enormous.jpg';
@@ -187,5 +188,32 @@ describe('buildSecondaryTags', () => {
       { id: ID2, uuid: UUID2, name: NAME2 },
       { id: ID1, uuid: UUID2, name: NAME1 },
     ]);
+  });
+});
+
+describe('.typeFrom', () => {
+  it('should return the correct type for an audio item', () => {
+    expect(typeFrom('moj_radio_item')).toBe('radio');
+  });
+  it('should return the correct type for an video item', () => {
+    expect(typeFrom('moj_video_item')).toBe('video');
+  });
+  it('should strip "node--" from the type string', () => {
+    expect(typeFrom('node--moj_video_item')).toBe('video');
+  });
+  it('should return the correct type for an pdf item', () => {
+    expect(typeFrom('moj_pdf_item')).toBe('pdf');
+  });
+
+  it('should return the correct type for a page', () => {
+    expect(typeFrom('page')).toBe('page');
+  });
+
+  it('should return the correct type for a series', () => {
+    expect(typeFrom('series')).toBe('series');
+  });
+
+  it('should return the correct type for a secondary tag', () => {
+    expect(typeFrom('tags')).toBe('tags');
   });
 });
