+++ server/utils/jsonApi.js
@@ -1,5 +1,3 @@
-const { typeFrom } = require('./adapters');
-
 const getImage = (data, type) => {
   if (!data) return null;
   return {
@@ -60,10 +58,25 @@ const buildSecondaryTags = arr =>
     name,
   }));
 
+const HUB_CONTENT_TYPES = {
+  moj_radio_item: 'radio',
+  moj_pdf_item: 'pdf',
+  moj_video_item: 'video',
+  page: 'page',
+  series: 'series',
+  tags: 'tags',
+};
+
+const typeFrom = type => {
+  const matches = type.match(/(?<=--)(.*)/g);
+  return HUB_CONTENT_TYPES[matches ? matches[0] : type];
+};
+
 module.exports = {
   getSmallTile,
   getLargeTile,
   getLargeImage,
   getCategoryIds,
   buildSecondaryTags,
+  typeFrom,
 };
